move Trade.tsx to components
