+++ client/src/components/Trade.tsx
